+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- **appSettings** query.
+- **cards.showMyCards** app setting.
+
 ## [0.28.1] - 2019-09-02
 
 ### Fixed